Second changes applied to code; rehashed database changes

+++ Interfaces/ILibraryBranchService.cs
@@ -10,7 +10,7 @@ namespace LibraryManagementSystem.Interfaces
         Task<LibraryBranch> Get(int branchId);
         Task<bool> UpdatePatronCount(int branchId, int patronNumber);
         Task<bool> UpdateAssetsCount(int branchId, int newAssets);
-        Task<bool> UpdateAssetAmount(int branchId, int newAmount);
+        Task<bool> UpdateAssetAmount(int branchId, decimal newAmount);
 
         Task<int> GetBooksCount(int branchId);
         Task<int> GetPatronNumber(int branchId);
